requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
torch==2.4.1
torchdata==0.7.1
dgl==1.1.3
pyyaml>=6.0
pydantic>=2.0.0